[3/n][vm-rewrite][Move] Update native context extensions to support Cloning, and let extension determine how to handle interior mutability.
This updates the `NativeContextExtensions` struct to be `Clone`able.
This updates the `Box<dyn Tid<'a>>` to a `Rc<dyn Tid<'a>>` and the
extension is responsible for handling interior mutability.

As a helper struct the `NativeContextMut` was added to allow for easy
writing/handling of native extensions that need interior mutability.

NB: The code in the PR may not be working as future PRs will build on
top of this. In particular, for this PR, the changes to the object
runtime/sui-adapter/sui-execution natives will be needed.