[4/n][vm-rewrite][Move] Add shared "old vm types" to move-execution, and update execution versions to use it
This gets the different execution versions working. It does this by
replicating old code and traits that are removed in the new VM into the
`move-execution/shared` location, and different move-execution and
sui-execution versioned code can access these things from there.

**NOTE:** This will be moved to a crate outside of `move-execution` in a
future PR. Changing the location of the new `shared` dir here will be
harder to do rather than making this change higher up.

NB: The code in the PR may not be working as future PRs will build on top of
this.