[6/n][vm-rewrite][sui-execution] Update object runtime to use type tags instead of runtime VM types.
Since the VM instance no longer will live across commands/the entire
lifetime of the object runtime, this replaces all VM runtime `Type`s
with `TypeTag`s or `MoveObjectType`s where appropriate. This also
updates the natives and runtime to handle the new
`NativeContextExtensions` model.

NB: The code in the PR may not be working as future PRs will build on top of
this.