[9/n][vm-rewrite][Move] Update depth calculation calculation/caching to be threadsafe
Switching the depth calculation of types to be lazily evaluated when we
first encountered the value, coupled with shared package cache, and the
types living in this shared package cache meant that the `depth` for a
type needed to be made threadsafe since we may need to write to it in a
parallel context.

This does this by `RwLock`ing the `depth` field and then making the
changes from there.